231004 commit

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -50,6 +50,13 @@ public class UserController {
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+
+        String readSql = "select * from user where name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "delete from user where name = ?";
         jdbcTemplate.update(sql, name);
     }
